caching_sha2_password: Do not send passwords over non-SSL connections
When server requests full authentication, ensure we only send the password
if the connection is over SSL, otherwise the password is in the plain and
authentication anyhow fails.

This only impacts users that explicitly set ssl_disabled=True, a
plain password on the wire is still very surprising given the RSA
infrastructure as a fallback that other clients use.